Add speech recognition with deepspeech, and a streaming recognizer.
The next step here is to improve the chunking of audio samples; currently
using a fixed 2s cadence and it's really choppy / drops lots of words. need
to have a more sensible silence detection algorithm

Brought deepspeech-rs in tree as a submodule to use my fork (updated to latest deepspeech)

Need to run with LD_LIBRARY_PATH=deepspeech-lib to get it to link. I don't know
what the right way is to get the build binary to find these libs (rpath?).
Of course in "real life" we'd expect the library to be installed into a system
root, but I don't want to do that (no libs in my package manager).