Added a VAD (voice activity detection) filter on speech being fed into deepspeech
It's a very basic thresholder based on levels I get on my current mic on full volume. I guess
that should be a bit more dynamic - maybe something in RustAudio can provide a less ad-hoc
way to write filters like this.

To assist with debugging, added functionality to play back clips at same time as sending them for recognition

To improve recognition results (with decent results) now use a back buffer, so that when
we hit the activation threshold we include whatever came earlier (to avoid chopping off
the start of speech segments)